test: Add tighten-bounds
In the spirit of the new big-{little,empty}-{box,triangle} tests, this test
combines various paths. However, these paths are not only filled but also used
for clipping, resulting in 120 different combinations.

No backend currently succeeds the test. The reference image is a gimp-ination of
the image and test-fallback results and thus certainly wrong. Feel free to fix.

Additionally, this makes the xcb backend die with an ugly failed assert.

Signed-off-by: Uli Schlachter <psychon@znc.in>