image: Try performing span composition a row at a time
In theory, this should be more cache efficient and allow us to trim the
operation to the width of row, shaving a few texel fetches. The cost is
that we cause pixman to evaluate the composite operation per-row. This
should only be a temporary solution until we can do something better
through pixman...

On a i5-2520m, ymmv,

  firefox-fishtank  64585.38 -> 56823.41:  1.14x speedup
  swfdec-fill-rate   1383.24  -> 1665.88:  1.20x slowdown

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>